requirements.txt
qrcode[pil]
pyzbar
opencv-python
numpy
matplotlib
pillow
streamlit